+++ src/AppCms.vue
@@ -1,5 +1,5 @@
 <template lang="pug">
-  #app
+  #AppCms
     router-view
 </template>
 
@@ -10,11 +10,5 @@ export default {
 </script>
 
 <style lang="stylus">
-#app
-  font-family 'Avenir', Helvetica, Arial, sans-serif
-  -webkit-font-smoothing antialiased
-  -moz-osx-font-smoothing grayscale
-  text-align center
-  color #2c3e50
-  margin-top 60px
+// #AppCms
 </style>