+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased](../../compare/master...dev)
 ### Added
+- Added component `GridComponents` to render dynamic components with the grid
 ### Changed
 - Removed CSS from base app component
 ### Deprecated