+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## [Unreleased](../../compare/master...dev)
 ### Added
 ### Changed
+- Updated vuejs webpack template from 1.2.7 to 1.2.8
 ### Deprecated
 ### Removed
 ### Fixed