+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 ### Changed
 - Updated vuejs webpack template from 1.2.7 to 1.2.8
+- Removed CSS from base app component
 ### Deprecated
 ### Removed
 ### Fixed