+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Add store page config
 - Add mock API for pages with random data components
 - Added component `GridComponents` to render dynamic components with the grid
+- Added wrapper for axios to keep http lib client isolated
+- Add PageApi
+- Load page componnets from mocked api on url change
 ### Changed
 - Updated vuejs webpack template from 1.2.7 to 1.2.8
 - Removed CSS from base app component