+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased](../../compare/master...dev)
 ### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+
+---
+## [0.6.0](../../compare/0.5.0...0.6.0) - 03/01/2018
+### Added
 - Add store page config
 - Add mock API for pages with random data components
 - Added component `GridComponents` to render dynamic components with the grid
@@ -17,12 +27,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Removed CSS from base app component
 - Removed demo components from `HelloWorld`and updated tests
 - Rename HelloWorld to PageRender
-### Deprecated
 ### Removed
 - Removed demo store counter and test
-### Fixed
-### Security
 
+---
 ## [0.5.0](../../compare/0.4.0...0.5.0) - 02/01/2018
 ### Added
 - Added /docs/ folder with build files